requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
arrow==0.15.7             # via -r requirements.in
asgiref==3.2.10           # via django
certifi==2020.6.20        # via requests
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
cryptography==2.9.2       # via social-auth-core
defusedxml==0.6.0         # via odfpy, python3-openid, social-auth-core
diff-match-patch==20181111  # via django-import-export
dj-database-url==0.5.0    # via -r requirements.in
django-admin-rangefilter==0.6.1  # via -r requirements.in
django-appconf==1.0.4     # via django-compressor
django-autocomplete-light==3.5.1  # via -r requirements.in
django-common-helpers==0.9.2  # via django-cron
django-compressor==2.4    # via -r requirements.in, django-libsass
django-cors-headers==3.4.0  # via -r requirements.in
django-cron==0.5.1        # via -r requirements.in
django-filter==2.3.0      # via -r requirements.in
django-import-export==2.2.0  # via -r requirements.in
django-libsass==0.8       # via -r requirements.in
django-model-utils==4.0.0  # via -r requirements.in
django-tz-detect==0.3.0   # via -r requirements.in
django-widget-tweaks==1.4.8  # via -r requirements.in
django==3.0.7             # via -r requirements.in, django-appconf, django-common-helpers, django-cors-headers, django-cron, django-filter, django-import-export, django-model-utils, django-tz-detect, djangorestframework, djangorestframework-simplejwt
django_pgviews==0.5.0     # via -r requirements.in
djangorestframework-camel-case==1.2.0  # via -r requirements.in
djangorestframework-csv==2.1.0  # via -r requirements.in
djangorestframework-gis==0.15  # via -r requirements.in
djangorestframework-simplejwt==4.4.0  # via -r requirements.in
djangorestframework==3.11.0  # via -r requirements.in, djangorestframework-csv, djangorestframework-gis, djangorestframework-simplejwt
et-xmlfile==1.0.1         # via openpyxl
gunicorn==20.0.4          # via -r requirements.in
idna==2.10                # via requests
inflection==0.5.0         # via -r requirements.in
jdcal==1.4.1              # via openpyxl
libsass==0.20.0           # via django-libsass
markuppy==1.14            # via tablib
oauthlib==3.1.0           # via requests-oauthlib, social-auth-core
odfpy==1.4.1              # via tablib
openpyxl==3.0.4           # via tablib
pillow==7.1.2             # via -r requirements.in
psycopg2==2.8.5           # via -r requirements.in
pycparser==2.20           # via cffi
pydash==4.8.0             # via -r requirements.in
pyjwt==1.7.1              # via djangorestframework-simplejwt, social-auth-core
python-dateutil==2.8.1    # via -r requirements.in, arrow
python3-openid==3.2.0     # via social-auth-core
pytz==2020.1              # via -r requirements.in, django, django-tz-detect
pyyaml==3.10              # via -r requirements.in, tablib
rcssmin==1.0.6            # via django-compressor
requests-oauthlib==1.3.0  # via social-auth-core
requests==2.24.0          # via -r requirements.in, requests-oauthlib, social-auth-core
rjsmin==1.1.0             # via django-compressor
six==1.15.0               # via cryptography, django-autocomplete-light, django-compressor, django-tz-detect, djangorestframework-csv, libsass, python-dateutil, social-auth-app-django, social-auth-core
social-auth-app-django==4.0.0  # via -r requirements.in
social-auth-core==3.3.3   # via social-auth-app-django
sqlparse==0.3.1           # via django
tablib[html,ods,xls,xlsx,yaml]==2.0.0  # via django-import-export
unicodecsv==0.14.1        # via djangorestframework-csv
urllib3==1.25.9           # via requests
uwsgi==2.0.19.1           # via -r requirements.in
xlrd==1.2.0               # via tablib
xlwt==1.3.0               # via tablib
xvfbwrapper==0.2.9        # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
